+++ Tests/AdHocTests/MVarTests.fs
@@ -8,6 +8,8 @@ open System.Numerics
 open Hopac
 open Hopac.Infixes
 
+exception Known
+
 let run () =
   let m = MVar 1
 
@@ -18,3 +20,8 @@ let run () =
   let n = MVar 0
 
   MVar.mutateFun ((+) 1) m ^->. 0 <|> MVar.modifyFun (fun n -> (n-1, n-2)) n |> run |> testEq -2
+
+  MVar.tryMutateFun ((+) 1) m ^->. 0 <|> MVar.tryModifyFun (fun n -> (n-1, n-2)) n |> run |> testEq -3
+
+  MVar.tryModifyFun (fun _ -> raise Known) n |> Job.catch <*> MVar.read n
+  |> run |> testEq (Choice2Of2 Known, -2)